requirements.txt
discord.py==0.16.12
boto3==1.9.97
